adapted to moved AbstractSimulationConfig
